+++ changelog.md
@@ -8,6 +8,7 @@
   - Mapper: switch image-rendering to crisp-edges instead of pixelated
 - **Fixed:**
   - Auto updater was updated to latest in attempt to fix auto update bugs.
+  - Mapper: Fixed an issue when adding a new room and not checking if a room exist at the same locations for an area
 
 ## 0.5.0 2018-08-25